Move J3DShapeMtx to correct header for assert
