+++ src/JSystem/J3DGraphBase/J3DPacket.cpp
@@ -2,6 +2,7 @@
 #include "JSystem/J3DGraphAnimator/J3DModel.h"
 #include "JSystem/J3DGraphBase/J3DDrawBuffer.h"
 #include "JSystem/J3DGraphBase/J3DMaterial.h"
+#include "JSystem/J3DGraphBase/J3DShapeMtx.h"
 #include "JSystem/JKernel/JKRHeap.h"
 #include <dolphin/os.h>
 #include <dolphin/os.h>
